+++ CHANGELOG.md
@@ -1,16 +1,18 @@
 # Diskover Change Log
 
-## [1.5.0-rc17] = 2018-10-01
+## [1.5.0-rc17] = 2018-10-02
 ### added
 - reduced crawl times
 - reduced number of es bulk updates and optimized frequency of bulk updates
 - improved crawl performance over nfs/cifs mounts
 - bots will now enqueue paths into redis queue (rq) if other bots are idle to improve crawl efficiency
+- improved adaptive batch algorithm to help speed up crawls, use with -a flag
 ### changed
 - removed python strftime which was slowing down crawls
 - removed filethreadtime from diskover.cfg.sample, removed thread code for long running directories
 - removed treethreads from diskover.cfg.sample, removed thread code for crawling directories in rootdir since
 provided no real benefit and was causing slow crawls over nfs and cifs
+- adaptive batch algorithm (-a)
 
 ## [1.5.0-rc16] = 2018-09-13
 ### fixed